pyproject.toml
[tool.poetry]
name = "jcli"
version = "0.0.2"
description = "Jcli: list, run, and check jenkins jobs"
authors = ["Brokenpip3 <brokenpip3@gmail.com>"]
license = "MIT"
repository = "https://github.com/brokenpip3/jcli"
readme = "README.md"

[tool.poetry.scripts]
jcli = "jcli.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
api4jenkins = "^1.9"
rich = "^12.0.0"
click = "^8.0.4"
appdirs = "^1.4.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
mock = "^4.0.3"
pytest = "^7.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
